[ES-001] feat/start-setup-env
