+++ README.md
@@ -7,7 +7,7 @@
 
 - `staging` — ветка для интеграции всех фич и изменений в тестовом окружении Здесь ведется тестировани;
 
-- `develop` — ветка для интеграции всех фич и изменений. Здесь ведется основная разработка;
+- `development` — ветка для интеграции всех фич и изменений. Здесь ведется основная разработка;
 
 - `feature/` — ветки для разработки конкретных задач или новых функций. Они создаются из develop и сливаются обратно после завершения работы;
 
@@ -52,31 +52,39 @@
     - через гитхаб
   - восстановление
     - ввод данных + новый пароль + повтор 
-    - подтверждение по почте SMTP
-основной функционал стартовой страницы:
-  - home (разные новости и т.п.)
-  - сайдменю (только для авторизованных)
-  - хеадер с различными элементами
-  - main
-  - footer
+  - подтверждение по почте SMTP
+  - подтверждение по СМС
+  -
 
 страницы:
   - для авторизованного:
-    - home
     - profile
     - friends
     - messages
-    - chat
-    - favorite
-    - settings
-  - без авторизации
     - video
     - music
     - games
-    - cripta
-    - maps
+    - chat
+    - settings
+  - без авторизации
+    - home
+    - video
+    - music
+    - users
+      - user
+    - stock-exchange
+    - map
     - weather
     - store
     - news
-
+    - posts
+    - survey
+  - страница ошибки
 ---
+
+основной функционал стартовой страницы:
+- home (разные новости и т.п.)
+- сайдменю (только для авторизованных)
+- хеадер с различными элементами
+- main
+- footer
