+++ express/package.json
@@ -1,43 +1,68 @@
 {
-  "name": "express",
-  "version": "1.0.0",
-  "main": "index.js",
-  "keywords": [],
+  "name": "social-sandbox-express",
+  "private": true,
+  "version": "0.0.1",
+  "main": "index.ts",
   "author": "Oleg Veremyev",
-  "type": "module",
   "contributors": [
     {
       "name": "Oleg Veremyev",
-      "email": "olegweremey1994@mail.ru"
+      "email": "olegweremey1994@mail.ru",
+      "url": "https://github.com/OlegWeremyew"
+    },
+    {
+      "name": "Artem Savchuk",
+      "email": "savchuk799@gmail.com",
+      "url": "https://github.com/RockStar666by"
     }
   ],
   "license": "ISC",
+  "keywords": [],
   "description": "",
   "scripts": {
-    "start": "node index.js",
-    "start:dev": "NODE_ENV=development nodemon index.js",
-    "start:prod": "NODE_ENV=production nodemon index.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "devDependencies": {
-    "@types/express": "^5.0.0",
-    "@types/mongoose": "^5.11.97",
-    "@types/node": "^22.13.5",
-    "@types/tailwindcss": "^3.1.0",
-    "nodemon": "^3.1.9",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.7.3"
+    "watcher": "tsc -w",
+    "start": "node dist/index.js",
+    "start:dev": "NODE_ENV=development nodemon dist/index.js",
+    "start:prod": "NODE_ENV=production nodemon dist/index.js",
+    "build:dev": "tsc -b && yarn build --mode development",
+    "build:stage": "tsc -b && yarn build --mode staging",
+    "build:prod": "tsc -b && yarn build --mode production",
+    "clean": "rm -rf node_modules && yarn cache clean",
+    "reset": "rm -rf node_modules && rm yarn.lock && yarn cache clean",
+    "reset_and_install": "rm -rf node_modules && rm yarn.lock && yarn cache clean && yarn"
   },
   "dependencies": {
+    "bcrypt": "^5.1.1",
     "body-parser": "^1.20.3",
     "cors": "^2.8.5",
     "dotenv": "^16.4.7",
     "express": "^4.21.2",
-    "install": "^0.13.0",
+    "express-validator": "^7.2.1",
+    "jsonwebtoken": "^9.0.2",
+    "mongodb": "^6.14.1",
     "mongoose": "^8.10.1",
+    "uuid": "^11.1.0"
+  },
+  "devDependencies": {
+    "@types/bcrypt": "^5.0.2",
+    "@types/cors": "^2.8.17",
+    "@types/express": "^5.0.0",
+    "@types/express-validator": "^3.0.2",
+    "@types/jsonwebtoken": "^9.0.9",
+    "@types/mongodb": "^4.0.7",
+    "@types/mongoose": "^5.11.97",
+    "@types/node": "^22.13.5",
+    "@types/uuid": "^10.0.0",
+    "nodemon": "^3.1.9",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.7.3"
   },
   "engines": {
     "node": ">=20",
     "yarn": ">=1.22"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/OlegWeremyew/sandbox-social/tree/production/express"
   }
 }
