+++ nest/package.json
@@ -9,7 +9,13 @@
   "contributors": [
     {
       "name": "Oleg Veremyev",
-      "email": "olegweremey1994@mail.ru"
+      "email": "olegweremey1994@mail.ru",
+      "url": "https://github.com/OlegWeremyew"
+    },
+    {
+      "name": "Artem Savchuk",
+      "email": "savchuk799@gmail.com",
+      "url": "https://github.com/RockStar666by"
     }
   ],
   "scripts": {
@@ -24,7 +30,10 @@
     "test:watch": "jest --watch",
     "test:cov": "jest --coverage",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
-    "test:e2e": "jest --config ./test/jest-e2e.json"
+    "test:e2e": "jest --config ./test/jest-e2e.json",
+    "clean": "rm -rf node_modules && yarn cache clean",
+    "reset": "rm -rf node_modules && rm yarn.lock && yarn cache clean",
+    "reset_and_install": "rm -rf node_modules && rm yarn.lock && yarn cache clean && yarn"
   },
   "dependencies": {
     "@nestjs/common": "^11.0.1",
@@ -80,5 +89,9 @@
   "engines": {
     "node": ">=20",
     "yarn": ">=1.22"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/OlegWeremyew/sandbox-social/tree/production/nest"
   }
 }
